requirements.txt

pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
pyyaml==6.0.2
fastapi==0.112.2
uvicorn==0.30.6
pyarrow==17.0.0
streamlit==1.39.0
matplotlib==3.9.2
